integrahub-api/requirements.txt
fastapi
uvicorn[standard]
pyjwt
python-dotenv
httpx
